Move crate structs to state and update code accordingly
