+++ pyth-sdk-solana/README.md
@@ -1,4 +1,4 @@
-# Pyth Client
+# Pyth SDK Solana
 
 This crate provides utilities for reading price feeds from the [pyth.network](https://pyth.network/) oracle on the Solana network.
 The crate includes a library for on-chain programs and an off-chain example program.
@@ -18,17 +18,17 @@ Add a dependency to your Cargo.toml:
 
 ```toml
 [dependencies]
-pyth-client="<version>"
+pyth-sdk-solana="<version>"
 ```
 
 If you want to use this library in your on-chain program you should use `no-entrypoint` feature to prevent conflict between your program and this library's program.
 
 ```toml
 [dependencies]
-pyth-client = {version = "<version>", features = ["no-entrypoint"]}
+pyth-sdk-solana = {version = "<version>", features = ["no-entrypoint"]}
 ```
 
-See [pyth-client on crates.io](https://crates.io/crates/pyth-client/) to get the latest version of the library.
+See [pyth-sdk-solana on crates.io](https://crates.io/crates/pyth-sdk-solana/) to get the latest version of the library.
 
 ## Usage
 
@@ -36,33 +36,19 @@ Pyth Network stores its price feeds in a collection of Solana accounts.
 This crate provides utilities for interpreting and manipulating the content of these accounts.
 Applications can obtain the content of these accounts in two different ways:
 * On-chain programs should pass these accounts to the instructions that require price feeds.
-* Off-chain programs can access these accounts using the Solana RPC client (as in the [example program](examples/get_accounts.rs)).
+* Off-chain programs can access these accounts using the Solana RPC client (as in the [example program](examples/eth_price.rs)).
 
 In both cases, the content of the account will be provided to the application as a binary blob (`Vec<u8>`).
 The examples below assume that the user has already obtained this account data.
 
 ### Parse account data
-
-Pyth Network has several different types of accounts:
-* Price accounts store the current price for a product
-* Product accounts store metadata about a product, such as its symbol (e.g., "BTC/USD").
-* Mapping accounts store a listing of all Pyth accounts
-
-For more information on the different types of Pyth accounts, see the [account structure documentation](https://docs.pyth.network/how-pyth-works/account-structure).
-The pyth.network website also lists the public keys of the accounts (e.g., [BTC/USD accounts](https://pyth.network/markets/#BTC/USD)).  
-
-This library provides several `load_*` methods that translate the binary data in each account into an appropriate struct: 
+To parse price account this library provides a `load_price` method which translates the binary data to Pyth Price structure. Pyth Price contains some functions to help working with the price. For more information please visit documentations of this crate.
 
 ```rust
-// replace with account data, either passed to on-chain program or from RPC node 
+use pyth_sdk_solana::{load_price, Price};
+
 let price_account_data: Vec<u8> = ...;
-let price_account: Price = load_price( &price_account_data ).unwrap();
-
-let product_account_data: Vec<u8> = ...;
-let product_account: Product = load_product( &product_account_data ).unwrap();
-
-let mapping_account_data: Vec<u8> = ...;
-let mapping_account: Mapping = load_mapping( &mapping_account_data ).unwrap();
+let price: Price = load_price( &price_account_data ).unwrap();
 ```
 
 ### Get the current price
@@ -77,12 +63,6 @@ println!("price: ({} +- {}) x 10^{}", price.price, price.conf, price.expo);
 The price is returned along with a confidence interval that represents the degree of uncertainty in the price.
 Both values are represented as fixed-point numbers, `a * 10^e`. 
 The method will return `None` if the price is not currently available.
-
-The status of the price feed determines if the price is available. You can get the current status using:
-
-```rust
-let price_status: PriceStatus = price_account.get_current_price_status();
-```
 
 ### Non-USD prices 
 
@@ -116,6 +96,34 @@ println!("0.1 BTC and 0.05 ETH are worth: ({} +- {}) x 10^{} USD",
 ```
 
 This function additionally propagates any uncertainty in the price into uncertainty in the value of the basket.
+
+### Solana Account Structure
+> It is adviced to use `load_price` above wherever possible as the account structure is designed specific to the Pyth onchain oracle and it is likely to change.
+
+Pyth Network has several different types of accounts:
+* Price accounts store the current price for a product
+* Product accounts store metadata about a product, such as its symbol (e.g., "BTC/USD").
+* Mapping accounts store a listing of all Pyth accounts
+
+For more information on the different types of Pyth accounts, see the [account structure documentation](https://docs.pyth.network/how-pyth-works/account-structure).
+The pyth.network website also lists the public keys of the accounts (e.g., [BTC/USD accounts](https://pyth.network/markets/#BTC/USD)).  
+
+This library provides several `load_*` methods in state module that translate the binary data in each account into an appropriate struct: 
+
+```rust
+use pyth_sdk_solana::state::*;
+
+// replace with account data, either passed to on-chain program or from RPC node 
+let price_account_data: Vec<u8> = ...;
+let price_account: &PriceAccount = load_price_account( &price_account_data ).unwrap();
+
+let product_account_data: Vec<u8> = ...;
+let product_account: &ProductAccount = load_product_account( &product_account_data ).unwrap();
+
+let mapping_account_data: Vec<u8> = ...;
+let mapping_account: &MappingAccount = load_mapping_account( &mapping_account_data ).unwrap();
+```
+
 
 ### Off-chain example program
