+++ pyth-sdk-solana/tests/common.rs
@@ -1,5 +1,5 @@
-use pyth_client::id;
-use pyth_client::processor::process_instruction;
+use pyth_sdk_solana::id;
+use pyth_sdk_solana::processor::process_instruction;
 use solana_program::instruction::Instruction;
 use solana_program_test::*;
 use solana_sdk::signature::Signer;
@@ -8,7 +8,7 @@ use solana_sdk::transaction::Transaction;
 // Panics if running instruction fails
 pub async fn test_instr_exec_ok(instr: Instruction) {
     let (mut banks_client, payer, recent_blockhash) =
-        ProgramTest::new("pyth_client", id(), processor!(process_instruction))
+        ProgramTest::new("pyth_sdk_solana", id(), processor!(process_instruction))
             .start()
             .await;
     let mut transaction = Transaction::new_with_payer(&[instr], Some(&payer.pubkey()));
